requirements.txt
absl-py==2.2.2
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
attrs==25.3.0
boto3==1.38.1
botocore==1.38.1
certifi==2025.1.31
cffi==1.17.1
channels==4.2.2
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
Django==5.2
django-seed==0.3.1
django-storages==1.14.6
djangorestframework==3.16.0
exceptiongroup==1.2.2
Faker==37.1.0
flatbuffers==25.2.10
fonttools==4.57.0
h11==0.14.0
httptools==0.6.4
idna==3.10
inflection==0.5.1
jax==0.6.0
jaxlib==0.6.0
jmespath==1.0.1
kiwisolver==1.4.8
matplotlib==3.10.1
mediapipe==0.10.21
minio==7.2.15
ml_dtypes==0.5.1
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
packaging==25.0
pillow==11.2.1
protobuf==4.25.6
psycopg2==2.9.10
pycparser==2.22
pycryptodome==3.22.0
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.12.0
scipy==1.15.2
sentencepiece==0.2.0
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.1
sqlparse==0.5.3
toposort==1.10
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
watchfiles==1.0.5
websockets==15.0.1
